+++ api/candidates.go
@@ -6,7 +6,17 @@ import (
 )
 
 func GetCandidates(w http.ResponseWriter, r *http.Request) {
-	cState := GetStateForRequest(r)
+	cState, err := GetStateForRequest(r)
+
+	if err != nil {
+		w.Header().Set("Content-Type", "application/json; charset=UTF-8")
+		w.WriteHeader(http.StatusNotFound)
+		_ = json.NewEncoder(w).Encode(Response{
+			Code: 404,
+			Log:  "State for given height not found",
+		})
+		return
+	}
 
 	candidates := cState.GetStateCandidates().GetData()
 
@@ -15,7 +25,7 @@ func GetCandidates(w http.ResponseWriter, r *http.Request) {
 	var result []Candidate
 
 	for _, candidate := range candidates {
-		result = append(result, makeResponseCandidate(candidate))
+		result = append(result, makeResponseCandidate(candidate, false))
 	}
 
 	w.WriteHeader(http.StatusOK)