+++ api/coin_info.go
@@ -16,8 +16,17 @@ type CoinInfoResponse struct {
 }
 
 func GetCoinInfo(w http.ResponseWriter, r *http.Request) {
+	cState, err := GetStateForRequest(r)
 
-	cState := GetStateForRequest(r)
+	if err != nil {
+		w.Header().Set("Content-Type", "application/json; charset=UTF-8")
+		w.WriteHeader(http.StatusNotFound)
+		_ = json.NewEncoder(w).Encode(Response{
+			Code: 404,
+			Log:  "State for given height not found",
+		})
+		return
+	}
 
 	vars := mux.Vars(r)
 	symbol := vars["symbol"]