+++ api/estimate_tx_commission.go
@@ -11,8 +11,17 @@ import (
 )
 
 func EstimateTxCommission(w http.ResponseWriter, r *http.Request) {
+	cState, err := GetStateForRequest(r)
 
-	cState := GetStateForRequest(r)
+	if err != nil {
+		w.Header().Set("Content-Type", "application/json; charset=UTF-8")
+		w.WriteHeader(http.StatusNotFound)
+		_ = json.NewEncoder(w).Encode(Response{
+			Code: 404,
+			Log:  "State for given height not found",
+		})
+		return
+	}
 
 	query := r.URL.Query()
 	rawTx := query.Get("tx")