+++ config/toml.go
@@ -102,6 +102,9 @@ db_path = "{{ js .BaseConfig.DBPath }}"
 # Output level for logging, including package level options
 log_level = "{{ .BaseConfig.LogLevel }}"
 
+# Path to file for logs, "stdout" by default
+log_path = "{{ .BaseConfig.LogPath }}"
+
 ##### additional base config options #####
 
 # Path to the JSON file containing the private key to use as a validator in the consensus protocol
@@ -228,4 +231,7 @@ prometheus_listen_addr = "{{ .Instrumentation.PrometheusListenAddr }}"
 # you increase your OS limits.
 # 0 - unlimited.
 max_open_connections = {{ .Instrumentation.MaxOpenConnections }}
+
+# Instrumentation namespace
+namespace = "minter"
 `