+++ CHANGELOG.md
@@ -11,6 +11,7 @@ BREAKING CHANGES
 IMPROVEMENT
 
 - [core] Add `MultisendTx`
+- [core] Add special cases to Formulas #140
 
 ## 0.6.0
 *Oct 30th, 2018*