+++ CHANGELOG.md
@@ -12,6 +12,7 @@ IMPROVEMENT
 
 - [core] Add `MultisendTx`
 - [core] Add special cases to Formulas #140
+- [core] Stake unbond now instant after dropping of from 1,000st place #146
 
 ## 0.6.0
 *Oct 30th, 2018*