+++ CHANGELOG.md
@@ -14,6 +14,7 @@ IMPROVEMENT
 - [core] Add special cases to Formulas #140
 - [core] Stake unbond now instant after dropping of from 1,000st place #146
 - [p2p] Default send and receive rates are now 15mB/s
+- [mempool] Set max mempool size to 10,000txs
 
 ## 0.6.0
 *Oct 30th, 2018*